split enums, interfaces and types into different files for organization
