+++ front-end/src/App.tsx
@@ -1,7 +1,8 @@
 import { useEffect } from "react";
-import { ThemeContextType } from "./context/ThemeContext";
-import { ThemeState, useTheme } from "./Hooks/useTheme"
+import { useTheme } from "./Hooks/useTheme"
 import { customFetch, requestTypeEnum } from "./lib/customFetch";
+import { ThemeState } from "./interfacesEnumsAndTypes/enums";
+import { ThemeContextType } from "./interfacesEnumsAndTypes/types";
 
 function App() {
   const { theme } : ThemeContextType = useTheme();
@@ -13,6 +14,7 @@ function App() {
       email: 'aaronwan808@gmail.com',
       password: 'test'
     }
+
     const jsonBody = JSON.stringify(body);
     const resp = await customFetch('/api/auth/login', requestTypeEnum.POST, jsonBody);
     return resp;
@@ -22,7 +24,7 @@ function App() {
     const resp = fetchThingy()
     console.log(resp);
   }, [])
-  
+
   return (
     <div className={`${theme === ThemeState.DARK ? "bg-black" : "bg-white"} min-h-screen w-full h-full`}>App</div>
   )
