+++ front-end/src/interfacesEnumsAndTypes/enums.ts
@@ -1,0 +1,11 @@
+export enum requestTypeEnum {
+    GET,
+    POST,
+    PUT,
+    DELETE
+};
+
+export enum ThemeState {
+  DARK,
+  LIGHT
+}
