+++ front-end/src/interfacesEnumsAndTypes/types.ts
@@ -1,0 +1,7 @@
+import { ThemeState } from "./enums";
+
+export type ThemeContextType = {
+  theme: ThemeState;
+  setTheme: React.Dispatch<React.SetStateAction<ThemeState>>;
+  toggleTheme: () => void;
+};
