+++ front-end/src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.tsx'
 import { ThemeProvider } from './context/ThemeContext.tsx'
-import { ThemeState } from './Hooks/useTheme.tsx'
+import { ThemeState } from './interfacesEnumsAndTypes/enums.ts'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
